sorting all the occurences
